Hotfix - Added authors - Moved dependencies related to development to dev - Refactored src to gatorconfig
