+++ gatorconfig/gui/check_file.py
