+++ pyproject.toml
@@ -1,8 +1,8 @@
 [tool.poetry]
-name = "project-template"
+name = "GatorConfig"
 version = "0.1.0"
-description = "Simple Python project template"
-authors = [""]
+description = "Autogeneration of GatorGradle configuration files."
+authors = ["Daniel Ullrich", "Kobe Coleman", "Paige Downey", "Favour Ojo"]
 license = "MIT"
 
 [tool.poetry.dependencies]
@@ -10,18 +10,19 @@ python = ">=3.7,<4.0"
 typer = "^0.4.0"
 "ruamel.yaml" = "^0.17.20"
 PyQt5 = "^5.15.6"
-pytest-qt = "^4.0.2"
-pytest-xvfb = "^2.0.0"
-mock = "^4.0.3"
 
 [tool.poetry.dev-dependencies]
 taskipy = "^1.9.0"
 pytest = "^6.2.5"
 pytest-cov = "^3.0.0"
 pylint = "^2.12.2"
+mock = "^4.0.3"
+pytest-mock = "^3.6.1"
+pytest-qt = "^4.0.2"
+pytest-xvfb = "^2.0.0"
 
 [tool.taskipy.tasks]
-test = "pytest --cov-report term-missing --cov-fail-under=70  --cov-branch --cov=src tests/ --ignore src/gui/"
+test = "pytest --cov-report term-missing --cov-fail-under=70  --cov-branch --cov=src tests/"
 lint = "pylint src/ tests/ src/gui/"
 
 [build-system]
