+++ tests/test_gator_yaml.py
@@ -1,6 +1,6 @@
 """Testing for gator_yaml"""
 import pytest
-from src.gator_yaml import GatorYaml
+from gatorconfig.gator_yaml import GatorYaml
 
 
 @pytest.mark.parametrize(
@@ -105,11 +105,11 @@ def test_enum_file_list(list_in, expected):
 @pytest.mark.parametrize(
     "files,expected",
     [
-        ({'src': {'main': {'java': {'test.java': ['']}}}},
-         "src:\n    main:\n        java:\n        "
+        ({'gatorconfig': {'main': {'java': {'test.java': ['']}}}},
+         "gatorconfig:\n    main:\n        java:\n        "
          "    test.java:\n                \n"),
-        ({'src': {'main': {'java': {'test.java': [''], 'test2.java': ['']}}}},
-         "src:\n    main:\n        java:\n            "
+        ({'gatorconfig': {'main': {'java': {'test.java': [''], 'test2.java': ['']}}}},
+         "gatorconfig:\n    main:\n        java:\n            "
          "test.java:\n                \n            test2.java:\n       "
          "         \n"),
     ],
